fix: auth.js 리팩토링
