+++ src/components/myProfile/ProfileImageForm.jsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo, useState } from 'react';
 import { toast } from 'react-toastify';
 import styled from 'styled-components';
-import { updateProfileImage } from '../../api/profile';
+import { updateProfileImage } from '../../api/profileApi';
 
 const ProfileImageForm = ({ profileImage, setProfileImage }) => {
   const [file, setFile] = useState(null);
@@ -21,7 +21,6 @@ const ProfileImageForm = ({ profileImage, setProfileImage }) => {
     async (e) => {
       e.preventDefault();
       if (!file) return toast.warning('업로드할 파일을 선택하세요!');
-
       const imageUrl = await updateProfileImage(file);
       setProfileImage(imageUrl);
       toast.success('프로필 사진이 변경되었습니다.');
@@ -29,6 +28,7 @@ const ProfileImageForm = ({ profileImage, setProfileImage }) => {
     [file, setProfileImage]
   );
 
+  // 
   const userProfileImage = useMemo(() => preview || profileImage, [preview, profileImage]);
 
   return (
