+++ src/pages/MyProfilePage.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import ProfileImageForm from '../components/myProfile/ProfileImageForm';
 import ProfileDataForm from '../components/myProfile/ProfileDataForm';
-import { fetchDefaultImage, fetchUserProfile } from '../api/profile';
+import { fetchDefaultImage, fetchUserProfile } from '../api/profileApi';
 
 const MyProfilePage = () => {
   const [profileImage, setProfileImage] = useState(null);
@@ -22,7 +22,6 @@ const MyProfilePage = () => {
         setCurrentNickname('Guest');
       }
     };
-
     fetchUserProfileImage();
   }, []);
